backend/requirements.txt
# ── Core framework ─────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.9.1

# ── Database ───────────────────────────────────────────
sqlalchemy[asyncio]==2.0.32
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ── Google Sheets API ──────────────────────────────────
gspread==6.1.2
google-auth==2.34.0

# ── Environment / config ──────────────────────────────
python-dotenv==1.0.1

# ── ML (Phase 3 — pre-install for convenience) ────────
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4

# ── Dev / testing ─────────────────────────────────────
httpx==0.27.0
pytest==8.3.2
pytest-asyncio==0.24.0
